% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_allowed}
\alias{validate_allowed}
\title{Validate allowed values.}
\usage{
validate_allowed(values, allowed, ids)
}
\arguments{
\item{values}{Vector of values}

\item{allowed}{Vector of allowed values}

\item{ids}{Vector of ids}
}
\value{
Errors data frame
}
\description{
Validate allowed values.
}
\keyword{internal}

