% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_any}
\alias{validate_any}
\title{Validate values with specific test.}
\usage{
validate_any(values, test, criteria, ids)
}
\arguments{
\item{values}{Vector of values}

\item{test}{Test name}

\item{criteria}{Test criteria}

\item{ids}{Vector of ids}
}
\value{
Errors data frame
}
\description{
Validate values with specific test.
}
\keyword{internal}

