% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_type}
\alias{validate_type}
\title{Validate type.}
\usage{
validate_type(values, type, ids)
}
\arguments{
\item{values}{Vector of values}

\item{type}{Allowed data type}

\item{ids}{Vector of ids}
}
\value{
Errors data frame
}
\description{
Validate type.
}
\keyword{internal}

